+++ CHANGES.rst
@@ -22,7 +22,7 @@ Features added
 * #13497: Support C domain objects in the table of contents.
 * #13535: html search: Update to the latest version of Snowball (v3.0.1).
   Patch by Adam Turner.
-* #13704: Autodoc: Detect :py:func:`typing_extensions.overload <typing.overload>`
+* #13704: autodoc: Detect :py:func:`typing_extensions.overload <typing.overload>`
   and :py:func:`~typing.final` decorators.
   Patch by Spencer Brown.